+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+- Google Analytics code for site analysis [#191](https://github.com/IN-CORE/incore-ui/issues/191)
 
 ## [1.11.0] - 2024-04-30